requirements.txt
Django>=4.1.1, <4.2
djangorestframework>=3.13.1, <3.14
drf-spectacular>=0.24.0, <0.25
psycopg2>=2.9.3, <2.10
requests>=2.28.1, <2.29